Pause Menu with 2/5 Buttons Working
Added pause menu that turns on and off when pressed with mouse, or when escape is pressed. Added ui for the buttons and screen. 2 buttons; return to game and return to main menu work. Functions like the game time pausing and audio need to be added before i can finish the other buttons.