Infobox and Shop Variables
Shop now displays the instance variables created when the game starts. Infobox is created with placeholder values. When you hover over a the first icon the infobox will display. Variables for infobox will use the variables from the monkey classes. Infobox will be finished when the class information is added.